Re-order integration tests
